+++ kompendium-core/src/main/kotlin/org/leafygreens/kompendium/annotations/HeaderParam.kt
@@ -2,4 +2,4 @@ package org.leafygreens.kompendium.annotations
 
 @Retention(AnnotationRetention.RUNTIME)
 @Target(AnnotationTarget.PROPERTY)
-annotation class HeaderParam
+annotation class HeaderParam(val description: String = "")