+++ kompendium-core/src/main/kotlin/org/leafygreens/kompendium/annotations/PathParam.kt
@@ -2,4 +2,4 @@ package org.leafygreens.kompendium.annotations
 
 @Retention(AnnotationRetention.RUNTIME)
 @Target(AnnotationTarget.PROPERTY)
-annotation class PathParam
+annotation class PathParam(val description: String = "")